+++ src/Client_Abstract.h
@@ -45,7 +45,7 @@ public:
 	virtual bool result(bool &success, uint32_t &status_code, uint64_t &file_size, std::wstring &content_type, std::wstring &content_encd) = 0;
 
 	//Read payload
-	virtual bool read_data(uint8_t *out_buff, const size_t &buff_size, size_t &bytes_read, bool &eof_flag) = 0;
+	virtual bool read_data(uint8_t *out_buff, const uint32_t &buff_size, size_t &bytes_read, bool &eof_flag) = 0;
 
 protected:
 	//WinINet initialization
